Refacto CseValidHandler to split responsabilities
