+++ gridcapa-cse-valid-app/src/test/java/com/farao_community/farao/cse_valid/app/CseValidHandlerTest.java
@@ -6,36 +6,19 @@
  */
 package com.farao_community.farao.cse_valid.app;
 
-import com.farao_community.farao.cse_valid.api.exception.CseValidInvalidDataException;
 import com.farao_community.farao.cse_valid.api.resource.CseValidFileResource;
 import com.farao_community.farao.cse_valid.api.resource.CseValidRequest;
 import com.farao_community.farao.cse_valid.api.resource.CseValidResponse;
 import com.farao_community.farao.cse_valid.api.resource.ProcessType;
 import com.farao_community.farao.cse_valid.app.configuration.EicCodesConfiguration;
-import com.farao_community.farao.cse_valid.app.dichotomy.DichotomyRunner;
-import com.farao_community.farao.cse_valid.app.exception.CseValidRequestValidatorException;
-import com.farao_community.farao.cse_valid.app.helper.LimitingElementHelper;
-import com.farao_community.farao.cse_valid.app.helper.NetPositionHelper;
 import com.farao_community.farao.cse_valid.app.mapper.EicCodesMapper;
-import com.farao_community.farao.cse_valid.app.rao.CseValidRaoValidator;
-import com.farao_community.farao.cse_valid.app.ttc_adjustment.TLimitingElement;
+import com.farao_community.farao.cse_valid.app.service.ComputeExportCornerService;
+import com.farao_community.farao.cse_valid.app.service.ComputeFullImportService;
 import com.farao_community.farao.cse_valid.app.ttc_adjustment.TTimestamp;
 import com.farao_community.farao.cse_valid.app.utils.CseValidRequestTestData;
 import com.farao_community.farao.cse_valid.app.utils.TimestampTestData;
-import com.farao_community.farao.cse_valid.app.validator.CseValidRequestValidator;
-import com.farao_community.farao.data.crac_api.Crac;
-import com.farao_community.farao.data.crac_creation.creator.cse.CseCracCreationContext;
-import com.farao_community.farao.data.rao_result_api.RaoResult;
-import com.farao_community.farao.dichotomy.api.results.DichotomyResult;
-import com.farao_community.farao.dichotomy.api.results.DichotomyStepResult;
 import com.farao_community.farao.minio_adapter.starter.MinioAdapter;
-import com.farao_community.farao.rao_runner.api.resource.RaoResponse;
-import com.powsybl.iidm.network.Network;
-import com.powsybl.iidm.network.VariantManager;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.mockito.MockedStatic;
-import org.mockito.Mockito;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -47,12 +30,9 @@ import java.time.ZoneOffset;
 import java.util.Objects;
 
 import static com.farao_community.farao.cse_valid.app.Constants.ERROR_MSG_CONTRADICTORY_DATA;
-import static com.farao_community.farao.cse_valid.app.Constants.ERROR_MSG_MISSING_CALCULATION_DIRECTIONS;
 import static com.farao_community.farao.cse_valid.app.Constants.ERROR_MSG_MISSING_DATA;
-import static com.farao_community.farao.cse_valid.app.Constants.ERROR_MSG_MISSING_SHIFTING_FACTORS;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
@@ -71,9 +51,6 @@ class CseValidHandlerTest {
     private MinioAdapter minioAdapter;
 
     @MockBean
-    private DichotomyRunner dichotomyRunner;
-
-    @MockBean
     private FileImporter fileImporter;
 
     @MockBean
@@ -83,10 +60,10 @@ class CseValidHandlerTest {
     private Logger businessLogger;
 
     @MockBean
-    private CseValidNetworkShifter cseValidNetworkShifter;
+    private ComputeFullImportService computeFullImportService;
 
     @MockBean
-    private CseValidRaoValidator cseValidRaoValidator;
+    private ComputeExportCornerService computeExportCornerService;
 
     @Autowired
     private EicCodesConfiguration eicCodesConfiguration;
@@ -147,7 +124,6 @@ class CseValidHandlerTest {
                 new CseValidFileResource("glsk.xml", "file://glsk.xml"),
                 OffsetDateTime.of(2020, 8, 12, 22, 30, 0, 0, ZoneOffset.UTC));
         when(minioAdapter.fileExists(any())).thenReturn(true);
-        when(dichotomyRunner.runImportCornerDichotomy(any(), any(), any(), any(), any())).thenReturn(null);
         CseValidResponse cseValidResponse = cseValidHandler.handleCseValidRequest(cseValidRequest);
         assertEquals("id", cseValidResponse.getId());
     }
@@ -156,7 +132,7 @@ class CseValidHandlerTest {
         return new CseValidFileResource(filename, Objects.requireNonNull(getClass().getResource("/" + filename)).toExternalForm());
     }
 
-    /* --------------- FULL IMPORT --------------- */
+    /* ------------------- computeTimestamp ------------------- */
 
     @Test
     void computeTimestampDataMissingMniiMnieMiec() {
@@ -211,6 +187,32 @@ class CseValidHandlerTest {
     }
 
     @Test
+    void computeTimestampMnii() {
+        CseValidRequest cseValidRequest = CseValidRequestTestData.getImportCseValidRequest(ProcessType.IDCC);
+        TTimestamp timestamp = TimestampTestData.getTimestampWithMnii();
+        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
+
+        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
+
+        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
+
+        verify(computeFullImportService, times(1)).computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
+    }
+
+    @Test
+    void computeTimestampMiec() {
+        CseValidRequest cseValidRequest = CseValidRequestTestData.getImportCseValidRequest(ProcessType.IDCC);
+        TTimestamp timestamp = TimestampTestData.getTimestampWithMiec();
+        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
+
+        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
+
+        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
+
+        verify(computeExportCornerService, times(1)).computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
+    }
+
+    @Test
     void computeTimestampMnie() {
         CseValidRequest cseValidRequest = CseValidRequestTestData.getImportCseValidRequest(ProcessType.IDCC);
         TTimestamp timestamp = TimestampTestData.getTimestampWithMnie();
@@ -221,591 +223,5 @@ class CseValidHandlerTest {
         cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
 
         verify(tcDocumentTypeWriter, times(1)).fillTimestampFullExportSuccess(timestamp, BigDecimal.ZERO);
-    }
-
-    @Test
-    void computeTimestampMniiMibniiAndAntcfinalAbsent() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getImportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMnii();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampFullImportSuccess(timestamp, BigDecimal.ZERO);
-    }
-
-    @Test
-    void computeTimestampMniiMibniiAndAntcfinalBothZero() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getImportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMibniiAndAntcfinalBothZero();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampFullImportSuccess(timestamp, BigDecimal.ZERO);
-    }
-
-    @Test
-    void computeTimestampMniiMibniiAbsent() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getImportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMniiAndAntcfinal();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampError(timestamp, ERROR_MSG_MISSING_DATA);
-    }
-
-    @Test
-    void computeTimestampMniiAntcfinalAbsent() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getImportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMniiAndMibnii();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampError(timestamp, ERROR_MSG_MISSING_DATA);
-    }
-
-    @Test
-    void computeTimestampMniiActualNtcAboveTarget() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getImportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMniiAndMibniiAndAntcfinalAndActualNtcAboveTarget();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(businessLogger, times(1)).info(anyString(), eq("time"));
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampFullImportSuccess(timestamp, BigDecimal.TEN);
-    }
-
-    @Test
-    void computeTimestampMniiFilesNotExisting() throws CseValidRequestValidatorException {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getImportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMniiAndMibniiAndAntcfinalAndActualNtcBelowTarget();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-        String errorMessage = "Process fail during TSO validation phase: Missing CGM file, CRAC file, GLSK file.";
-        CseValidRequestValidatorException e = new CseValidRequestValidatorException(errorMessage);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        try (MockedStatic<CseValidRequestValidator> cseValidRequestValidatorMockedStatic = Mockito.mockStatic(CseValidRequestValidator.class)) {
-            cseValidRequestValidatorMockedStatic.when(() -> CseValidRequestValidator.checkAllFilesExist(cseValidRequest, false))
-                    .thenThrow(e);
-            cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-        }
-
-        verify(businessLogger, times(1)).error(anyString(), eq("time"));
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampError(timestamp, e.getMessage());
-    }
-
-    @Test
-    void computeTimestampMniiRunDichotomyError() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getImportCseValidRequest(ProcessType.IDCC);
-        String cgmUrl = cseValidRequest.getCgm().getUrl();
-        String cracUrl = cseValidRequest.getImportCrac().getUrl();
-        ProcessType processType = cseValidRequest.getProcessType();
-        OffsetDateTime processTargetDateTime = cseValidRequest.getTimestamp();
-
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMniiAndMibniiAndAntcfinalAndActualNtcBelowTarget();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        String jsonCracUrl = "/CSE/VALID/crac.utc";
-        String raoParameterUrl = "/CSE/VALID/raoParameter.utc";
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-        Network network = mock(Network.class);
-        CseCracCreationContext cracCreationContext = mock(CseCracCreationContext.class);
-        Crac crac = mock(Crac.class);
-        when(cracCreationContext.getCrac()).thenReturn(crac);
-
-        when(fileImporter.importNetwork(cgmUrl)).thenReturn(network);
-        when(fileImporter.importCracCreationContext(cracUrl, processTargetDateTime, network)).thenReturn(cracCreationContext);
-        when(fileExporter.saveCracInJsonFormat(crac, processTargetDateTime, processType)).thenReturn(jsonCracUrl);
-        when(fileExporter.saveRaoParameters(processTargetDateTime, processType)).thenReturn(raoParameterUrl);
-
-        when(dichotomyRunner.runImportCornerDichotomy(timestampWrapper, cseValidRequest, jsonCracUrl, raoParameterUrl, network)).thenReturn(null);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillDichotomyError(timestamp);
-    }
-
-    @Test
-    void computeTimestampMniiRunDichotomySuccessHighestValidStepNotNull() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getImportCseValidRequest(ProcessType.IDCC);
-        String cgmUrl = cseValidRequest.getCgm().getUrl();
-        String cracUrl = cseValidRequest.getImportCrac().getUrl();
-        ProcessType processType = cseValidRequest.getProcessType();
-        OffsetDateTime processTargetDateTime = cseValidRequest.getTimestamp();
-
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMniiAndMibniiAndAntcfinalAndActualNtcBelowTarget();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        String jsonCracUrl = "/CSE/VALID/crac.utc";
-        String raoParameterUrl = "/CSE/VALID/raoParameter.utc";
-        String networkFileUrl = "CSE/Valid/network.utc";
-        String raoResultFileUrl = "CSE/VALID/raoResult.utc";
-        TLimitingElement limitingElement = new TLimitingElement();
-        double fullImportValue = 10.0;
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-        Network network = mock(Network.class);
-        CseCracCreationContext cracCreationContext = mock(CseCracCreationContext.class);
-        Crac crac = mock(Crac.class);
-        when(cracCreationContext.getCrac()).thenReturn(crac);
-        DichotomyResult<RaoResponse> dichotomyResult = mock(DichotomyResult.class);
-        RaoResponse raoResponse = mock(RaoResponse.class);
-        RaoResult raoResult = mock(RaoResult.class);
-        DichotomyStepResult<RaoResponse> highestValidStep = mock(DichotomyStepResult.class);
-
-        when(fileImporter.importNetwork(cgmUrl)).thenReturn(network);
-        when(fileImporter.importCracCreationContext(cracUrl, processTargetDateTime, network)).thenReturn(cracCreationContext);
-        when(fileImporter.importNetwork(networkFileUrl)).thenReturn(network);
-
-        when(fileExporter.saveCracInJsonFormat(crac, processTargetDateTime, processType)).thenReturn(jsonCracUrl);
-        when(fileExporter.saveRaoParameters(processTargetDateTime, processType)).thenReturn(raoParameterUrl);
-
-        when(dichotomyResult.hasValidStep()).thenReturn(true);
-        when(dichotomyResult.getHighestValidStep()).thenReturn(highestValidStep);
-        when(highestValidStep.getValidationData()).thenReturn(raoResponse);
-        when(fileImporter.importRaoResult(raoResultFileUrl, crac)).thenReturn(raoResult);
-        when(dichotomyRunner.runImportCornerDichotomy(timestampWrapper, cseValidRequest, jsonCracUrl, raoParameterUrl, network)).thenReturn(dichotomyResult);
-
-        when(raoResponse.getNetworkWithPraFileUrl()).thenReturn(networkFileUrl);
-        when(raoResponse.getRaoResultFileUrl()).thenReturn(raoResultFileUrl);
-        try (
-                MockedStatic<LimitingElementHelper> limitingElementServiceMockedStatic = Mockito.mockStatic(LimitingElementHelper.class);
-                MockedStatic<NetPositionHelper> netPositionMockedStatic = Mockito.mockStatic(NetPositionHelper.class)
-        ) {
-            limitingElementServiceMockedStatic.when(() -> LimitingElementHelper.getLimitingElement(raoResult, cracCreationContext, network))
-                    .thenReturn(limitingElement);
-            netPositionMockedStatic.when(() -> NetPositionHelper.computeItalianImport(network))
-                    .thenReturn(fullImportValue);
-            cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-        }
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampWithFullImportDichotomyResponse(timestamp, BigDecimal.ONE, fullImportValue, limitingElement);
-    }
-
-    /* --------------- EXPORT CORNER --------------- */
-
-    @Test
-    void computeTimestampMiecMibiecAndAntcfinalAbsent() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMiec();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampExportCornerSuccess(timestamp, BigDecimal.ZERO);
-    }
-
-    @Test
-    void computeTimestampMiecMibiecAndAntcfinalBothZero() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMibiecAndAntcfinalBothZero();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampExportCornerSuccess(timestamp, BigDecimal.ZERO);
-    }
-
-    @Test
-    void computeTimestampMiecMibiecAbsent() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMiecAndAntcfinal();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampError(timestamp, ERROR_MSG_MISSING_DATA);
-    }
-
-    @Test
-    void computeTimestampMiecAntcfinalAbsent() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMiecAndMibiec();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampError(timestamp, ERROR_MSG_MISSING_DATA);
-    }
-
-    @Test
-    void computeTimestampMiecShiftingFactorsMissing() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithoutShiftingFactors();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampError(timestamp, ERROR_MSG_MISSING_SHIFTING_FACTORS);
-    }
-
-    @Test
-    void computeTimestampMiecCalculationDirectionsMissing() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithoutCalculationDirections();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampError(timestamp, ERROR_MSG_MISSING_CALCULATION_DIRECTIONS);
-    }
-
-    @Test
-    void computeTimestampMiecActualNtcAboveTarget() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithMiecAndMibiecAndAntcfinalAndActualNtcAboveTarget();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampExportCornerSuccess(timestamp, BigDecimal.TEN);
-    }
-
-    @Test
-    void computeTimestampMiecWhithoutFranceInAreaOrOutAreaShouldThrowAnException() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithoutFranceInAreaOrOutArea();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        Assertions.assertThatExceptionOfType(CseValidInvalidDataException.class)
-                .isThrownBy(() -> cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter));
-    }
-
-    @Test
-    void computeTimestampMiecWithFranceInAreaAndFilesNotExisting() throws CseValidRequestValidatorException {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithFranceInArea();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        String errorMessage = "Process fail during TSO validation phase: Missing CGM file, CRAC file, GLSK file, CRAC Transit.";
-        CseValidRequestValidatorException e = new CseValidRequestValidatorException(errorMessage);
-
-        try (MockedStatic<CseValidRequestValidator> cseValidRequestValidatorMockedStatic = Mockito.mockStatic(CseValidRequestValidator.class)) {
-            cseValidRequestValidatorMockedStatic.when(() -> CseValidRequestValidator.checkAllFilesExist(cseValidRequest, true))
-                    .thenThrow(e);
-            cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-        }
-
-        verify(businessLogger, times(1)).error(anyString(), eq("time"));
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampError(timestamp, e.getMessage());
-    }
-
-    @Test
-    void computeTimestampMiecWithFranceOutAreaAndFilesNotExisting() throws CseValidRequestValidatorException {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        TTimestamp timestamp = TimestampTestData.getTimestampWithFranceOutArea();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-
-        String errorMessage = "Process fail during TSO validation phase: Missing CGM file, CRAC file, GLSK file.";
-        CseValidRequestValidatorException e = new CseValidRequestValidatorException(errorMessage);
-
-        try (MockedStatic<CseValidRequestValidator> cseValidRequestValidatorMockedStatic = Mockito.mockStatic(CseValidRequestValidator.class)) {
-            cseValidRequestValidatorMockedStatic.when(() -> CseValidRequestValidator.checkAllFilesExist(cseValidRequest, false))
-                    .thenThrow(e);
-            cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-        }
-
-        verify(businessLogger, times(1)).error(anyString(), eq("time"));
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampError(timestamp, e.getMessage());
-    }
-
-    @Test
-    void computeTimestampMiecWithFranceInAreaShouldNotRunDichotomyForExportCornerBecauseNetworkShiftedIsSecure() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        String cgmUrl = cseValidRequest.getCgm().getUrl();
-        String glskUrl = cseValidRequest.getGlsk().getUrl();
-        String cracUrl = cseValidRequest.getExportCrac().getUrl();
-        ProcessType processType = cseValidRequest.getProcessType();
-        OffsetDateTime processTargetDateTime = cseValidRequest.getTimestamp();
-
-        TTimestamp timestamp = TimestampTestData.getTimestampWithFranceInArea();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        String jsonCracUrl = "/CSE/VALID/crac.utc";
-        String raoParameterUrl = "/CSE/VALID/raoParameter.utc";
-        String basePath = "IDCC/2023/01/09/12_30/ARTIFACTS/";
-        String variantName = "1234";
-        String networkNameOrId = "test";
-        String scaledNetworkDirPath = basePath + variantName;
-        String networkFilePath = scaledNetworkDirPath + networkNameOrId + ".xiidm";
-        String networkFileUrl = "CSE/Valid/network.utc";
-        String resultsDestination = "CSE/VALID/" + scaledNetworkDirPath;
-        double shiftValue = timestampWrapper.getMiecIntValue() - (timestampWrapper.getMibiecIntValue() - timestampWrapper.getAntcfinalIntValue());
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-        Network network = mock(Network.class);
-        CseCracCreationContext cracCreationContext = mock(CseCracCreationContext.class);
-        Crac crac = mock(Crac.class);
-        when(cracCreationContext.getCrac()).thenReturn(crac);
-        VariantManager variantManager = mock(VariantManager.class);
-        RaoResponse raoResponse = mock(RaoResponse.class);
-
-        when(fileImporter.importNetwork(cgmUrl)).thenReturn(network);
-        when(fileImporter.importCracCreationContext(cracUrl, processTargetDateTime, network)).thenReturn(cracCreationContext);
-        when(fileExporter.saveCracInJsonFormat(crac, processTargetDateTime, processType)).thenReturn(jsonCracUrl);
-        when(fileExporter.saveRaoParameters(processTargetDateTime, processType)).thenReturn(raoParameterUrl);
-        when(fileExporter.makeDestinationMinioPath(processTargetDateTime, processType, FileExporter.FileKind.ARTIFACTS)).thenReturn(basePath);
-        when(fileExporter.saveNetworkInArtifact(network, networkFilePath, "", processTargetDateTime, processType)).thenReturn(networkFileUrl);
-
-        when(network.getNameOrId()).thenReturn(networkNameOrId);
-        when(network.getVariantManager()).thenReturn(variantManager);
-        when(variantManager.getWorkingVariantId()).thenReturn(variantName);
-
-        when(cseValidRaoValidator.runRao(cseValidRequest, networkFileUrl, jsonCracUrl, raoParameterUrl, resultsDestination)).thenReturn(raoResponse);
-        when(cseValidRaoValidator.isSecure(raoResponse)).thenReturn(true);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(cseValidNetworkShifter, times(1)).shiftNetwork(shiftValue, network, timestampWrapper, glskUrl);
-        verify(cseValidRaoValidator, times(1)).runRao(cseValidRequest, networkFileUrl, jsonCracUrl, raoParameterUrl, resultsDestination);
-        verify(cseValidRaoValidator, times(1)).isSecure(raoResponse);
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampExportCornerSuccess(timestamp, timestamp.getMIEC().getV());
-    }
-
-    @Test
-    void computeTimestampMiecWithFranceOutAreaShouldNotRunDichotomyForExportCornerBecauseNetworkShiftedIsSecure() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        String cgmUrl = cseValidRequest.getCgm().getUrl();
-        String glskUrl = cseValidRequest.getGlsk().getUrl();
-        String cracUrl = cseValidRequest.getImportCrac().getUrl();
-        ProcessType processType = cseValidRequest.getProcessType();
-        OffsetDateTime processTargetDateTime = cseValidRequest.getTimestamp();
-
-        TTimestamp timestamp = TimestampTestData.getTimestampWithFranceOutArea();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        String jsonCracUrl = "/CSE/VALID/crac.utc";
-        String raoParameterUrl = "/CSE/VALID/raoParameter.utc";
-        String basePath = "IDCC/2023/01/09/12_30/ARTIFACTS/";
-        String variantName = "1234";
-        String networkNameOrId = "test";
-        String scaledNetworkDirPath = basePath + variantName;
-        String networkFilePath = scaledNetworkDirPath + networkNameOrId + ".xiidm";
-        String networkFileUrl = "CSE/Valid/network.utc";
-        String resultsDestination = "CSE/VALID/" + scaledNetworkDirPath;
-        double shiftValue = timestampWrapper.getMiecIntValue() - (timestampWrapper.getMibiecIntValue() - timestampWrapper.getAntcfinalIntValue());
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-        Network network = mock(Network.class);
-        CseCracCreationContext cracCreationContext = mock(CseCracCreationContext.class);
-        Crac crac = mock(Crac.class);
-        when(cracCreationContext.getCrac()).thenReturn(crac);
-        VariantManager variantManager = mock(VariantManager.class);
-        RaoResponse raoResponse = mock(RaoResponse.class);
-
-        when(fileImporter.importNetwork(cgmUrl)).thenReturn(network);
-        when(fileImporter.importCracCreationContext(cracUrl, processTargetDateTime, network)).thenReturn(cracCreationContext);
-        when(fileExporter.saveCracInJsonFormat(crac, processTargetDateTime, processType)).thenReturn(jsonCracUrl);
-        when(fileExporter.saveRaoParameters(processTargetDateTime, processType)).thenReturn(raoParameterUrl);
-        when(fileExporter.makeDestinationMinioPath(processTargetDateTime, processType, FileExporter.FileKind.ARTIFACTS)).thenReturn(basePath);
-        when(fileExporter.saveNetworkInArtifact(network, networkFilePath, "", processTargetDateTime, processType)).thenReturn(networkFileUrl);
-
-        when(network.getNameOrId()).thenReturn(networkNameOrId);
-        when(network.getVariantManager()).thenReturn(variantManager);
-        when(variantManager.getWorkingVariantId()).thenReturn(variantName);
-
-        when(cseValidRaoValidator.runRao(cseValidRequest, networkFileUrl, jsonCracUrl, raoParameterUrl, resultsDestination)).thenReturn(raoResponse);
-        when(cseValidRaoValidator.isSecure(raoResponse)).thenReturn(true);
-
-        cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-
-        verify(cseValidNetworkShifter, times(1)).shiftNetwork(shiftValue, network, timestampWrapper, glskUrl);
-        verify(cseValidRaoValidator, times(1)).runRao(cseValidRequest, networkFileUrl, jsonCracUrl, raoParameterUrl, resultsDestination);
-        verify(cseValidRaoValidator, times(1)).isSecure(raoResponse);
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampExportCornerSuccess(timestamp, timestamp.getMIEC().getV());
-    }
-
-    @Test
-    void computeTimestampMiecWithFranceInAreaShouldRunDichotomyForExportCorner() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        String cgmUrl = cseValidRequest.getCgm().getUrl();
-        String glskUrl = cseValidRequest.getGlsk().getUrl();
-        String cracUrl = cseValidRequest.getExportCrac().getUrl();
-        ProcessType processType = cseValidRequest.getProcessType();
-        OffsetDateTime processTargetDateTime = cseValidRequest.getTimestamp();
-
-        TTimestamp timestamp = TimestampTestData.getTimestampWithFranceInArea();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TLimitingElement limitingElement = new TLimitingElement();
-
-        String jsonCracUrl = "/CSE/VALID/crac.utc";
-        String raoParametersUrl = "/CSE/VALID/raoParameter.utc";
-        String basePath = "IDCC/2023/01/09/12_30/ARTIFACTS/";
-        String variantName = "1234";
-        String networkNameOrId = "test";
-        String scaledNetworkDirPath = basePath + variantName;
-        String networkFilePath = scaledNetworkDirPath + networkNameOrId + ".xiidm";
-        String networkFileUrl = "CSE/Valid/network.utc";
-        String resultsDestination = "CSE/VALID/" + scaledNetworkDirPath;
-        String raoResultFileUrl = "CSE/VALID/raoResult.utc";
-        double shiftValue = timestampWrapper.getMiecIntValue() - (timestampWrapper.getMibiecIntValue() - timestampWrapper.getAntcfinalIntValue());
-        double exportCornerValue = 10.0;
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-        Network network = mock(Network.class);
-        CseCracCreationContext cracCreationContext = mock(CseCracCreationContext.class);
-        Crac crac = mock(Crac.class);
-        when(cracCreationContext.getCrac()).thenReturn(crac);
-        VariantManager variantManager = mock(VariantManager.class);
-        RaoResponse raoResponse = mock(RaoResponse.class);
-        DichotomyResult<RaoResponse> dichotomyResult = mock(DichotomyResult.class);
-        RaoResult raoResult = mock(RaoResult.class);
-        DichotomyStepResult<RaoResponse> highestValidStep = mock(DichotomyStepResult.class);
-
-        when(fileImporter.importNetwork(cgmUrl)).thenReturn(network);
-        when(fileImporter.importCracCreationContext(cracUrl, processTargetDateTime, network)).thenReturn(cracCreationContext);
-        when(fileImporter.importNetwork(networkFileUrl)).thenReturn(network);
-        when(fileImporter.importRaoResult(raoResultFileUrl, crac)).thenReturn(raoResult);
-
-        when(fileExporter.saveCracInJsonFormat(crac, processTargetDateTime, processType)).thenReturn(jsonCracUrl);
-        when(fileExporter.saveRaoParameters(processTargetDateTime, processType)).thenReturn(raoParametersUrl);
-        when(fileExporter.makeDestinationMinioPath(processTargetDateTime, processType, FileExporter.FileKind.ARTIFACTS)).thenReturn(basePath);
-        when(fileExporter.saveNetworkInArtifact(network, networkFilePath, "", processTargetDateTime, processType)).thenReturn(networkFileUrl);
-
-        when(network.getNameOrId()).thenReturn(networkNameOrId);
-        when(network.getVariantManager()).thenReturn(variantManager);
-        when(variantManager.getWorkingVariantId()).thenReturn(variantName);
-
-        when(cseValidRaoValidator.runRao(cseValidRequest, networkFileUrl, jsonCracUrl, raoParametersUrl, resultsDestination)).thenReturn(raoResponse);
-        when(cseValidRaoValidator.isSecure(raoResponse)).thenReturn(false);
-        when(raoResponse.getNetworkWithPraFileUrl()).thenReturn(networkFileUrl);
-        when(raoResponse.getRaoResultFileUrl()).thenReturn(raoResultFileUrl);
-
-        when(dichotomyRunner.runExportCornerDichotomy(timestampWrapper, cseValidRequest, jsonCracUrl, raoParametersUrl, network)).thenReturn(dichotomyResult);
-        when(dichotomyResult.hasValidStep()).thenReturn(true);
-        when(dichotomyResult.getHighestValidStep()).thenReturn(highestValidStep);
-        when(highestValidStep.getValidationData()).thenReturn(raoResponse);
-        try (
-                MockedStatic<LimitingElementHelper> limitingElementServiceMockedStatic = Mockito.mockStatic(LimitingElementHelper.class);
-                MockedStatic<NetPositionHelper> netPositionMockedStatic = Mockito.mockStatic(NetPositionHelper.class)
-        ) {
-            limitingElementServiceMockedStatic.when(() -> LimitingElementHelper.getLimitingElement(raoResult, cracCreationContext, network))
-                    .thenReturn(limitingElement);
-            netPositionMockedStatic.when(() -> NetPositionHelper.computeFranceImportFromItaly(network))
-                    .thenReturn(exportCornerValue);
-            cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-        }
-
-        verify(cseValidNetworkShifter, times(1)).shiftNetwork(shiftValue, network, timestampWrapper, glskUrl);
-        verify(cseValidRaoValidator, times(1)).runRao(cseValidRequest, networkFileUrl, jsonCracUrl, raoParametersUrl, resultsDestination);
-        verify(cseValidRaoValidator, times(1)).isSecure(raoResponse);
-        verify(dichotomyRunner, times(1)).runExportCornerDichotomy(timestampWrapper, cseValidRequest, jsonCracUrl, raoParametersUrl, network);
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampWithExportCornerDichotomyResponse(timestamp, limitingElement, exportCornerValue, true);
-    }
-
-    @Test
-    void computeTimestampMiecWithFranceOutAreaShouldRunDichotomyForExportCorner() {
-        CseValidRequest cseValidRequest = CseValidRequestTestData.getExportCseValidRequest(ProcessType.IDCC);
-        String cgmUrl = cseValidRequest.getCgm().getUrl();
-        String glskUrl = cseValidRequest.getGlsk().getUrl();
-        String cracUrl = cseValidRequest.getImportCrac().getUrl();
-        ProcessType processType = cseValidRequest.getProcessType();
-        OffsetDateTime processTargetDateTime = cseValidRequest.getTimestamp();
-
-        TTimestamp timestamp = TimestampTestData.getTimestampWithFranceOutArea();
-        TTimestampWrapper timestampWrapper = new TTimestampWrapper(timestamp, eicCodesConfiguration, eicCodesMapper);
-
-        TLimitingElement limitingElement = new TLimitingElement();
-
-        String jsonCracUrl = "/CSE/VALID/crac.utc";
-        String raoParametersUrl = "/CSE/VALID/raoParameter.utc";
-        String basePath = "IDCC/2023/01/09/12_30/ARTIFACTS/";
-        String variantName = "1234";
-        String networkNameOrId = "test";
-        String scaledNetworkDirPath = basePath + variantName;
-        String networkFilePath = scaledNetworkDirPath + networkNameOrId + ".xiidm";
-        String networkFileUrl = "CSE/Valid/network.utc";
-        String resultsDestination = "CSE/VALID/" + scaledNetworkDirPath;
-        String raoResultFileUrl = "CSE/VALID/raoResult.utc";
-        double shiftValue = timestampWrapper.getMiecIntValue() - (timestampWrapper.getMibiecIntValue() - timestampWrapper.getAntcfinalIntValue());
-        double exportCornerValue = 10.0;
-
-        TcDocumentTypeWriter tcDocumentTypeWriter = mock(TcDocumentTypeWriter.class);
-        Network network = mock(Network.class);
-        CseCracCreationContext cracCreationContext = mock(CseCracCreationContext.class);
-        Crac crac = mock(Crac.class);
-        when(cracCreationContext.getCrac()).thenReturn(crac);
-        VariantManager variantManager = mock(VariantManager.class);
-        RaoResponse raoResponse = mock(RaoResponse.class);
-        DichotomyResult<RaoResponse> dichotomyResult = mock(DichotomyResult.class);
-        RaoResult raoResult = mock(RaoResult.class);
-        DichotomyStepResult<RaoResponse> highestValidStep = mock(DichotomyStepResult.class);
-
-        when(fileImporter.importNetwork(cgmUrl)).thenReturn(network);
-        when(fileImporter.importCracCreationContext(cracUrl, processTargetDateTime, network)).thenReturn(cracCreationContext);
-        when(fileImporter.importNetwork(networkFileUrl)).thenReturn(network);
-        when(fileImporter.importRaoResult(raoResultFileUrl, crac)).thenReturn(raoResult);
-
-        when(fileExporter.saveCracInJsonFormat(crac, processTargetDateTime, processType)).thenReturn(jsonCracUrl);
-        when(fileExporter.saveRaoParameters(processTargetDateTime, processType)).thenReturn(raoParametersUrl);
-        when(fileExporter.makeDestinationMinioPath(processTargetDateTime, processType, FileExporter.FileKind.ARTIFACTS)).thenReturn(basePath);
-        when(fileExporter.saveNetworkInArtifact(network, networkFilePath, "", processTargetDateTime, processType)).thenReturn(networkFileUrl);
-
-        when(network.getNameOrId()).thenReturn(networkNameOrId);
-        when(network.getVariantManager()).thenReturn(variantManager);
-        when(variantManager.getWorkingVariantId()).thenReturn(variantName);
-
-        when(cseValidRaoValidator.runRao(cseValidRequest, networkFileUrl, jsonCracUrl, raoParametersUrl, resultsDestination)).thenReturn(raoResponse);
-        when(cseValidRaoValidator.isSecure(raoResponse)).thenReturn(false);
-        when(raoResponse.getNetworkWithPraFileUrl()).thenReturn(networkFileUrl);
-        when(raoResponse.getRaoResultFileUrl()).thenReturn(raoResultFileUrl);
-
-        when(dichotomyRunner.runExportCornerDichotomy(timestampWrapper, cseValidRequest, jsonCracUrl, raoParametersUrl, network)).thenReturn(dichotomyResult);
-        when(dichotomyResult.hasValidStep()).thenReturn(true);
-        when(dichotomyResult.getHighestValidStep()).thenReturn(highestValidStep);
-        when(highestValidStep.getValidationData()).thenReturn(raoResponse);
-        try (
-                MockedStatic<LimitingElementHelper> limitingElementServiceMockedStatic = Mockito.mockStatic(LimitingElementHelper.class);
-                MockedStatic<NetPositionHelper> netPositionMockedStatic = Mockito.mockStatic(NetPositionHelper.class)
-        ) {
-            limitingElementServiceMockedStatic.when(() -> LimitingElementHelper.getLimitingElement(raoResult, cracCreationContext, network))
-                    .thenReturn(limitingElement);
-            netPositionMockedStatic.when(() -> NetPositionHelper.computeFranceImportFromItaly(network))
-                    .thenReturn(exportCornerValue * -1);
-            cseValidHandler.computeTimestamp(timestampWrapper, cseValidRequest, tcDocumentTypeWriter);
-        }
-
-        verify(cseValidNetworkShifter, times(1)).shiftNetwork(shiftValue, network, timestampWrapper, glskUrl);
-        verify(cseValidRaoValidator, times(1)).runRao(cseValidRequest, networkFileUrl, jsonCracUrl, raoParametersUrl, resultsDestination);
-        verify(cseValidRaoValidator, times(1)).isSecure(raoResponse);
-        verify(dichotomyRunner, times(1)).runExportCornerDichotomy(timestampWrapper, cseValidRequest, jsonCracUrl, raoParametersUrl, network);
-        verify(tcDocumentTypeWriter, times(1)).fillTimestampWithExportCornerDichotomyResponse(timestamp, limitingElement, exportCornerValue, false);
     }
 }
